add group test
