+++ mux_test.go
@@ -138,27 +138,6 @@ func TestMuxParam3(t *testing.T) {
 	p.Do(t)
 }
 
-func TestMuxGroupParam(t *testing.T) {
-	r := NewRouter()
-
-	users := r.Group("/users/:name")
-	users.Get("/:age", func(w http.ResponseWriter, r *http.Request) {
-		w.Write([]byte("name=" + URLParam(r, "name") + ", age=" + URLParam(r, "age")))
-	})
-
-	p := &Pattern{
-		Reqests: []*Reqest{
-			{"/users/aaa/24", 200, "name=aaa, age=24"},
-			{"/users/bbb/23", 200, "name=bbb, age=23"},
-		},
-
-		Server: httptest.NewServer(r),
-	}
-
-	defer p.Close()
-	p.Do(t)
-}
-
 func TestMuxBackTrack(t *testing.T) {
 	r := NewRouter()
 
@@ -206,48 +185,6 @@ func TestMuxMiddleware(t *testing.T) {
 			{"/users/b", 200, "MMb"},
 		},
 
-		Server: httptest.NewServer(r),
-	}
-
-	defer p.Close()
-	p.Do(t)
-}
-
-func TestMuxGroupMiddleware(t *testing.T) {
-	r := NewRouter()
-
-	a := r.Group("/a")
-	a.Use(MiddlewareTest("MA"))
-	a.Get("/a", func(w http.ResponseWriter, r *http.Request) {
-		w.Write([]byte("a"))
-	})
-
-	aa := a.Group("/a")
-	aa.Use(MiddlewareTest("MAA"))
-	aa.Get("/a", func(w http.ResponseWriter, r *http.Request) {
-		w.Write([]byte("aa"))
-	})
-
-	r.Use(MiddlewareTest("M"))
-
-	b := r.Group("/b")
-	b.Use(MiddlewareTest("MB"))
-	b.Get("/b", func(w http.ResponseWriter, r *http.Request) {
-		w.Write([]byte("b"))
-	})
-
-	c := r.Group("/c")
-	c.Get("/c", func(w http.ResponseWriter, r *http.Request) {
-		w.Write([]byte("c"))
-	})
-
-	p := &Pattern{
-		Reqests: []*Reqest{
-			{"/a/a", 200, "MAa"},
-			{"/a/a/a", 200, "MAMAAaa"},
-			{"/b/b", 200, "MMBb"},
-			{"/c/c", 200, "Mc"},
-		},
 		Server: httptest.NewServer(r),
 	}
